Evil hack: B/c an ssize_t is 4 bytes on most platforms, but 8 bytes on dux, and b/c our checkpoint file protocol is so broken and doesn't use cedar, we want to use ints (always 4 bytes) everywhere.  That way we don't get stuck in an infinite loop with Alphas trying to read 8 bytes of file size for an ack while the other end only sent 4 bytes.

+++ src/condor_ckpt/image.C
@@ -685,9 +685,9 @@ int
 Image::Write( int fd )
 {
 	int		i;
-	ssize_t	pos = 0;
-	ssize_t	nbytes;
-	ssize_t ack;
+	int		pos = 0;
+	int		nbytes;
+	int		ack;
 	int		status;
 
 		// Write out the header
@@ -750,7 +750,7 @@ int
 Image::Read()
 {
 	int		i;
-	ssize_t	nbytes;
+	int		nbytes;
 
 		// Make sure we have a valid file descriptor to read from
 	if( fd < 0 && file_name && file_name[0] ) {
@@ -797,12 +797,12 @@ Image::Close()
 ssize_t
 SegMap::Read( int fd, ssize_t pos )
 {
-	ssize_t nbytes;
+	int		nbytes;
 	char *orig_brk;
 	char *cur_brk;
 	char	*ptr;
-	size_t	bytes_to_go;
-	size_t	read_size;
+	int		bytes_to_go;
+	int		read_size;
 	long	saved_len = len;
 	int 	saved_prot = prot;
 	RAW_ADDR	saved_core_loc = core_loc;
